requirements.txt
# Core LLM and KG
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-neo4j>=0.4.0
langchain-experimental>=0.3.0
langgraph>=0.2.0

# Neo4j driver
neo4j>=5.0.0

# Data models and validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Document processing
pymupdf>=1.24.0        # PDF ingestion (incremental, Sprint 2)
markdown-it-py>=3.0.0  # Markdown parsing

# RDF/OWL export (Sprint 2)
rdflib>=7.0.0

# Utilities
rich>=13.0.0
tqdm>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# App (Sprint 4)
streamlit>=1.35.0
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0          # FastAPI TestClient dependency

# Optional: LightRAG community retrieval
# lightrag-hku[neo4j]>=1.0.0
